$YOLO fix
Former-commit-id: 17a36301c9083dbedc2b886d9827c432cc6d8a45 [formerly 934bc0653c9d995061834de93c8b4afc567e06fc]
Former-commit-id: 85c37df644526b89eb86b56c539307a8761f20e8
Former-commit-id: b15554d6a04c199bc2df20aacfb68bbb1f897b28

+++ install-dotfiles.sh
@@ -52,7 +52,7 @@ NC='\033[0m'
 
 function confirm() {
     echo -en "$LBLUE |||$GREEN Do you want to override ${LGREEN}$1 $LBLUE(Y/n)? >> $NC"
-    if [ $YOLO -eq 1 ]; then
+    if [ -n $YOLO ] && [ $YOLO -eq 1 ]; then
         echo y
         rm -rf "$1"
         return
